requirements.txt
discord.py==2.2.2
python-dotenv==1.0.0
asgiref==3.6.0
openai==0.27.6
revChatGPT==5.0.2
selenium==4.9.1
undetected-chromedriver==3.4.7
flask
schedule
